requirements.txt
# 基本套件
requests==2.31.0
beautifulsoup4==4.12.2
pandas>=2.0.0
lxml>=4.9.0
python-dotenv==1.0.0
schedule==1.2.0
openpyxl>=3.1.0

# MongoDB套件
pymongo==4.6.0
dnspython==2.4.2

# Selenium套件（用於JavaScript渲染）
selenium==4.15.2
webdriver-manager==4.0.1

# 其他工具
fake-useragent==1.4.0
urllib3>=2.0.0
